moves files core rtkquery files into metabase/api
